Rx2: Follow Rx1 rules and move isViewReady to RxTiUtils
