+++ rx2/src/main/java/net/grandcentrix/thirtyinch/rx2/RxTiPresenterUtils.java
@@ -15,56 +15,17 @@
 
 package net.grandcentrix.thirtyinch.rx2;
 
-import net.grandcentrix.thirtyinch.Removable;
-import net.grandcentrix.thirtyinch.TiLifecycleObserver;
 import net.grandcentrix.thirtyinch.TiPresenter;
-import net.grandcentrix.thirtyinch.TiView;
 
 import io.reactivex.Observable;
-import io.reactivex.ObservableEmitter;
-import io.reactivex.ObservableOnSubscribe;
-import io.reactivex.disposables.Disposable;
 
 public class RxTiPresenterUtils {
 
     /**
-     * Observable of the view state. The View is ready to receive calls after calling {@link
-     * TiPresenter#attachView(TiView)} and before calling {@link TiPresenter#detachView()}.
+     * @deprecated use {@link RxTiUtils#isViewReady(TiPresenter)} instead
      */
     public static Observable<Boolean> isViewReady(final TiPresenter presenter) {
-        return Observable.create(new ObservableOnSubscribe<Boolean>() {
-            @Override
-            public void subscribe(final ObservableEmitter<Boolean> emitter)
-                    throws Exception {
-                if (!emitter.isDisposed()) {
-                    emitter.onNext(presenter.getState() == TiPresenter.State.VIEW_ATTACHED);
-                }
-
-                final Removable removable = presenter
-                        .addLifecycleObserver(new TiLifecycleObserver() {
-                            @Override
-                            public void onChange(final TiPresenter.State state,
-                                    final boolean hasLifecycleMethodBeenCalled) {
-                                if (!emitter.isDisposed()) {
-                                    emitter.onNext(state == TiPresenter.State.VIEW_ATTACHED
-                                            && hasLifecycleMethodBeenCalled);
-                                }
-                            }
-                        });
-
-                emitter.setDisposable(new Disposable() {
-                    @Override
-                    public void dispose() {
-                        removable.remove();
-                    }
-
-                    @Override
-                    public boolean isDisposed() {
-                        return removable.isRemoved();
-                    }
-                });
-            }
-        }).distinctUntilChanged();
+        return RxTiUtils.isViewReady(presenter);
     }
 
 }
